Update jquery.tokenize.js
Added width and singleColumn options.  Modified AJAX data to support additional key options and count, which the server can use to report back how many rows were found.  Test is made against maxElements and generates error message which is sent to erroMessageField option.  This is currently hardcoded, so it needs to be adjusted for flexibility.